Force the alpha to 1 if blending is disabled
It seems that WGL drops the color buffer's alpha on the floor. In WebGL,
it blends against the page. So the "white" materials we were seeing were
the page's background. I can wonder why the faceship has an output alpha
of 0 for these materials, but for now, just force alpha to 1.0 if
there's no blending.